Move `RegionServiceInterface` to `Region` folder
