+++ src/ElasticSearch/Offer/GeoShapeQueryRegionService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Search\ElasticSearch\Offer;
 
+use CultuurNet\UDB3\Search\ElasticSearch\Region\RegionServiceInterface;
 use CultuurNet\UDB3\Search\Region\RegionId;
 use Elasticsearch\Client;
